requirements.txt
Flask
Flask-Login
Flask-MySQLdb
requests
Werkzeug
firebase-admin
python-dotenv
